feat: add news section with 3-column layout using NewsAPI and Promise.allSettled
