--- scripts/constants/newsCategoryPalette.js
@@ -1,29 +1,0 @@
-// Categorías mostradas en el sidebar, con su etiqueta visible,
-// la categoría real de NewsAPI y la paleta de colores del diseño.
-// NewsAPI soporta: business, entertainment, general, health, science, sports, technology.
-export const NEWS_CATEGORIES = [
-  {
-    label: "TECNOLOGÍA",
-    apiCategory: "technology",
-    color: "#6842ff",
-    background: "#1e1444",
-  },
-  {
-    label: "DEPORTES",
-    apiCategory: "sports",
-    color: "#00d2ff",
-    background: "#003344",
-  },
-  {
-    label: "POLÍTICA",
-    apiCategory: "general",
-    color: "#06d6a0",
-    background: "#003320",
-  },
-  {
-    label: "CULTURA",
-    apiCategory: "entertainment",
-    color: "#ff2d78",
-    background: "#3d1428",
-  },
-];
